cluster mode

+++ handlers/instance_handlers.go
@@ -1,6 +1,8 @@
 package handlers
 
 import (
+	"fmt"
+
 	"gitee.com/flexlb/flexlb-api/config"
 	"gitee.com/flexlb/flexlb-api/restapi/operations/instance"
 
@@ -11,13 +13,22 @@ type InstanceCreateHandlerImpl struct {
 }
 
 func (h *InstanceCreateHandlerImpl) Handle(params instance.CreateParams) middleware.Responder {
+	// cluster instancce blackout
+	if params.Config.Name == config.LB.Cluster.Name {
+		msg := fmt.Sprintf("cluster instance '%s' cannot be created manually", params.Config.Name)
+		return instance.NewCreateBadRequest().WithPayload(&instance.CreateBadRequestBody{Message: &msg})
+	}
+
+	// create instance
 	inst, err := config.CreateInstance(params.Config)
 	if err != nil {
-		var errMsg instance.CreateBadRequestBody
 		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewCreateBadRequest().WithPayload(&errMsg)
+		return instance.NewCreateBadRequest().WithPayload(&instance.CreateBadRequestBody{Message: &msg})
 	}
+
+	// notify other nodes
+	config.GossipCreateInstance(inst)
+
 	return instance.NewCreateOK().WithPayload(inst)
 }
 
@@ -35,10 +46,8 @@ type InstanceGetHandlerImpl struct {
 func (h *InstanceGetHandlerImpl) Handle(params instance.GetParams) middleware.Responder {
 	inst, err := config.GetInstance(params.Name)
 	if err != nil {
-		var errMsg instance.GetBadRequestBody
 		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewGetBadRequest().WithPayload(&errMsg)
+		return instance.NewGetBadRequest().WithPayload(&instance.GetBadRequestBody{Message: &msg})
 	}
 	return instance.NewGetOK().WithPayload(inst)
 }
@@ -47,13 +56,22 @@ type InstanceModifyHandlerImpl struct {
 }
 
 func (h *InstanceModifyHandlerImpl) Handle(params instance.ModifyParams) middleware.Responder {
-	inst, err := config.ModifyInstance(params.Name, params.Config)
-	if err != nil {
-		var errMsg instance.ModifyBadRequestBody
-		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewModifyBadRequest().WithPayload(&errMsg)
+	// cluster instancce blackout
+	if params.Config.Name == config.LB.Cluster.Name {
+		msg := fmt.Sprintf("cluster instance '%s' cannot be modified manually", params.Config.Name)
+		return instance.NewModifyBadRequest().WithPayload(&instance.ModifyBadRequestBody{Message: &msg})
 	}
+
+	// modify instance
+	inst, err := config.ModifyInstance(params.Config)
+	if err != nil {
+		msg := err.Error()
+		return instance.NewModifyBadRequest().WithPayload(&instance.ModifyBadRequestBody{Message: &msg})
+	}
+
+	// notify other nodes
+	config.GossipModifyInstance(inst)
+
 	return instance.NewModifyOK().WithPayload(inst)
 }
 
@@ -61,13 +79,22 @@ type InstanceDeleteHandlerImpl struct {
 }
 
 func (h *InstanceDeleteHandlerImpl) Handle(params instance.DeleteParams) middleware.Responder {
+	// cluster instancce blackout
+	if params.Name == config.LB.Cluster.Name {
+		msg := fmt.Sprintf("cluster instance '%s' cannot be deleted manually", params.Name)
+		return instance.NewDeleteBadRequest().WithPayload(&instance.DeleteBadRequestBody{Message: &msg})
+	}
+
+	// delete instance
 	err := config.DeleteInstance(params.Name)
 	if err != nil {
-		var errMsg instance.DeleteBadRequestBody
 		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewDeleteBadRequest().WithPayload(&errMsg)
+		return instance.NewDeleteBadRequest().WithPayload(&instance.DeleteBadRequestBody{Message: &msg})
 	}
+
+	// notify other nodes
+	config.GossipDeleteInstance(params.Name)
+
 	return instance.NewDeleteOK()
 }
 
@@ -75,13 +102,22 @@ type InstanceStopHandlerImpl struct {
 }
 
 func (h *InstanceStopHandlerImpl) Handle(params instance.StopParams) middleware.Responder {
+	// cluster instancce blackout
+	if params.Name == config.LB.Cluster.Name {
+		msg := fmt.Sprintf("cluster instance '%s' cannot be stoped manually", params.Name)
+		return instance.NewStopBadRequest().WithPayload(&instance.StopBadRequestBody{Message: &msg})
+	}
+
+	// stop instance
 	inst, err := config.StopInstance(params.Name)
 	if err != nil {
-		var errMsg instance.StopBadRequestBody
 		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewStopBadRequest().WithPayload(&errMsg)
+		return instance.NewStopBadRequest().WithPayload(&instance.StopBadRequestBody{Message: &msg})
 	}
+
+	// notify other nodes
+	config.GossipStopInstance(params.Name)
+
 	return instance.NewStopOK().WithPayload(inst)
 }
 
@@ -89,12 +125,21 @@ type InstanceStartHandlerImpl struct {
 }
 
 func (h *InstanceStartHandlerImpl) Handle(params instance.StartParams) middleware.Responder {
+	// cluster instancce blackout
+	if params.Name == config.LB.Cluster.Name {
+		msg := fmt.Sprintf("cluster instance '%s' cannot be started manually", params.Name)
+		return instance.NewStartBadRequest().WithPayload(&instance.StartBadRequestBody{Message: &msg})
+	}
+
+	// start instance
 	inst, err := config.StartInstance(params.Name)
 	if err != nil {
-		var errMsg instance.StartBadRequestBody
 		msg := err.Error()
-		errMsg.Message = &msg
-		return instance.NewStartBadRequest().WithPayload(&errMsg)
+		return instance.NewStartBadRequest().WithPayload(&instance.StartBadRequestBody{Message: &msg})
 	}
+
+	// notify other nodes
+	config.GossipStartInstance(params.Name)
+
 	return instance.NewStartOK().WithPayload(inst)
 }
